LayoutTests/streams/reference-implementation/readable-stream-expected.txt


PASS ReadableStream can be constructed with no errors 
FAIL ReadableStream instances should have the correct list of properties assert_array_equals: should have all the correct methods lengths differ, expected 6 got 5
PASS ReadableStream constructor should throw for non-function start arguments 
PASS ReadableStream constructor can get initial garbage as cancel argument 
PASS ReadableStream constructor can get initial garbage as pull argument 
PASS ReadableStream constructor can get initial garbage as strategy argument 
TIMEOUT ReadableStream start should be able to return a promise Test timed out
TIMEOUT ReadableStream start should be able to return a promise and reject it Test timed out
PASS ReadableStream should be able to enqueue different objects. 
PASS ReadableStream: if start throws an error, it should be re-thrown 
TIMEOUT ReadableStream: if pull rejects, it should error the stream Test timed out
FAIL ReadableStream: should not call pull until the previous pull call's promise fulfills assert_equals: after the promise returned by pull is fulfilled, pull should be called a second time expected 2 but got 1
PASS ReadableStream: should not call pull after start if the stream is now closed 
FAIL ReadableStream: should call pull after enqueueing from inside pull (with no read requests), if strategy allows assert_equals: pull() should have been called four times expected 4 but got 1
PASS ReadableStream pull should be able to close a stream. 
FAIL ReadableStream: enqueue should throw when the stream is readable but draining assert_equals: the first enqueue should return true expected (boolean) true but got (undefined) undefined
PASS ReadableStream: enqueue should throw when the stream is closed 
PASS ReadableStream: enqueue should throw the stored error when the stream is errored 
FAIL ReadableStream: should call underlying source methods as methods releaseLock is not implemented
FAIL ReadableStream strategies: the default strategy should return false for all but the first enqueue call assert_equals: first enqueue should return true expected (boolean) true but got (undefined) undefined
FAIL ReadableStream strategies: the default strategy should continue returning true from enqueue if the chunks are read immediately assert_equals: first enqueue should return true expected (boolean) true but got (undefined) undefined
TIMEOUT ReadableStream integration test: adapting a random push source Test timed out
PASS ReadableStream integration test: adapting a sync pull source 
